array-*: rename exercises
