--- exercises/concept/array-transformations/.docs/after.md
@@ -1,3 +1,0 @@
-TODO: Add more.
-
-Remember, anytime you're about to write a for loop to transform an array there is a good chance Array has already has you covered with a built-in.
