+++ exercises/concept/elyses-analytic-enchantments/.eslintrc
