+++ exercises/concept/elyses-analytic-enchantments/.meta/config.json
@@ -11,7 +11,11 @@
     }
   ],
   "editor": {
-    "solution_files": ["array-analysis.js"],
-    "test_files": ["array-analysis.spec.js"]
+    "solution_files": [
+      "enchantments.js"
+    ],
+    "test_files": [
+      "enchantments.spec.js"
+    ]
   }
 }
