+++ exercises/concept/elyses-analytic-enchantments/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "@exercism/javascript-concept-array-destructuring",
-  "description": "Exercism concept exercise on array destructuring",
-  "author": "Kristina Born",
+  "name": "@exercism/javascript-elyses-analytic-enchantments",
+  "description": "Exercism concept exercise on array analysis",
+  "author": "Peterchu999",
   "version": "1.0.0",
   "license": "MIT",
   "private": true,
   "repository": {
     "type": "git",
     "url": "https://github.com/exercism/v3.git",
-    "directory": "languages/javascript/exercises/concept/array-destructuring"
+    "directory": "languages/javascript/exercises/concept/elyses-analytic-enchantments"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
