+++ exercises/concept/elyses-destructured-enchantments/global.d.ts
