+++ exercises/concept/elyses-destructured-enchantments/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "@exercism/javascript-concept-arrays",
-  "description": "Exercism concept exercise on arrays",
-  "author": "Ovidiu Miu",
+  "name": "@exercism/javascript-elyses-destructured-enchantments",
+  "description": "Exercism concept exercise on array destructuring",
+  "author": "Kristina Born",
   "version": "1.0.0",
   "license": "MIT",
   "private": true,
   "repository": {
     "type": "git",
     "url": "https://github.com/exercism/v3.git",
-    "directory": "languages/javascript/exercises/concept/arrays"
+    "directory": "languages/javascript/exercises/concept/elyses-destructured-enchantments"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
